// package/src/types.ts
export type FileItem = {
  id: string;
  name: string;
  type: 'file' | 'folder';
  parent_id: string | null;
  employee_id?: string;
  size?: string;
  date: string;
  extension?: string;
  url?: string;
  password?: string;
};